+++ pos_chat/static/src/xml/chat.xml
@@ -1,6 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates id="template" xml:space="preserve">
 
+    <t t-name="ChatButton">
+      <div class="control-button">
+        <i class="fa fa-list-alt" />
+          Chat Button
+      </div>
+    </t>
+
     <t t-name="CustomScreenWidget">
       <div class="custom-screen screen">
         <div class="screen-content">
@@ -15,21 +22,17 @@
              </span>
            </section>
            <section class="full-content">
-             <div class="window">
-               <section class="subwindow collapsed">
-                 <div class="subwindow-container collapsed">
-                   <div class="subwindow-container-fix custom-details-contents" />
-                 </div>
-               </section>
-             </div>
+               <div class="chat_window">
+                   <section class="whit-space">
+
+                   </section>
+                   <section class="input-text">
+
+                   </section>
+               </div>
            </section>
         </div>
       </div>
     </t>
 
-    <t t-name="PopupButton">
-      <div class="control-button">
-        <i class="fa fa-list-alt" /> Popup Button
-      </div>
-    </t>
 </templates>
