+++ pos_chat/view/chat_rooms.xml
@@ -3,7 +3,7 @@
 
     <template id="assets" inherit_id="point_of_sale.assets">
           <xpath expr="." position="inside">
-              <script type="text/javascript" src="/pos_chat/static/src/js/pos.js"></script>
+              <script type="text/javascript" src="/pos_chat/static/src/js/chat_button.js"></script>
           </xpath>
     </template>
